Upgrade bash if necessary
Ran into an issue where readline install will fail if an older version of bash is being run (see https://github.com/Homebrew/homebrew-core/issues/5799). A simple `brew upgrade bash` fixed it. This is to have the bootstrapper upgrade it if necessary to prevent others from running into it if they have an older bash as well on Mac.